server/requirements.txt

bidict==0.21.2
gunicorn==20.0.4
python-engineio==4.2.1
python-socketio==4.6.0